Try to find a python interpreter which is python 2
"python" is apparently python 3 on some systems, and the scripts do
not run in python 3.  Thus, try to find the name of a python 2
interpreter when installing the scripts.  (Currently it doesn't
actually check the python version, but assumes that an interpreter
named "python2" or "python2.x" should be preferred to one which
is simply named "python".)